✒️Rename: mail 기능 domain 분리
