+++ src/main/kotlin/com/example/days/domain/mail/dto/request/EmailRequest.kt
@@ -1,4 +1,4 @@
-package com.example.days.domain.user.dto.request
+package com.example.days.domain.mail.dto.request
 
 data class EmailRequest(
     val email: String
