+++ src/main/kotlin/com/example/days/domain/mail/repository/MailRepository.kt
@@ -1,6 +1,6 @@
-package com.example.days.domain.user.repository
+package com.example.days.domain.mail.repository
 
-import com.example.days.domain.user.model.Mail
+import com.example.days.domain.mail.model.Mail
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface MailRepository: JpaRepository<Mail, Long> {
