+++ src/main/kotlin/com/example/days/domain/mail/service/MailService.kt
@@ -1,6 +1,6 @@
-package com.example.days.domain.user.service
+package com.example.days.domain.mail.service
 
-import com.example.days.domain.user.dto.request.EmailRequest
+import com.example.days.domain.mail.dto.request.EmailRequest
 
 interface UserMailService {
     fun sendVerificationEmail(request: EmailRequest)
