+++ src/main/kotlin/com/example/days/domain/user/service/UserServiceImpl.kt
@@ -1,7 +1,8 @@
 package com.example.days.domain.user.service
 
+import com.example.days.domain.mail.dto.request.EmailRequest
 import com.example.days.domain.user.dto.request.*
-import com.example.days.domain.user.dto.response.EmailResponse
+import com.example.days.domain.mail.dto.response.EmailResponse
 import com.example.days.domain.user.dto.response.LoginResponse
 import com.example.days.domain.user.dto.response.ModifyInfoResponse
 import com.example.days.domain.user.dto.response.SignUpResponse
